requirements.txt
anyio==4.0.0
beautifulsoup4==4.12.2
browser-cookie3==0.19.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
deep-translator==1.11.4
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-translate==3.12.1
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.1
hyperframe==6.0.1
idna==3.4
jeepney==0.8.0
langdetect==1.0.9
lz4==4.3.2
objsize==0.7.0
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.19.0
python-dotenv==1.0.0
python-telegram-bot==20.6
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
urllib3==2.1.0
